update 8/8 import excel

+++ lib/model/eqc_quote/model_input_quote_detail.dart
@@ -9,8 +9,8 @@ class InputQuoteDetail {
   String? damageDetail;
   int? quantity;
   String? dimension;
-  int? length;
-  int? width;
+  double? length;
+  double? width;
   String? location;
   double? laborCost;
   double? mrCost;
